Popover menu: move under explore/loggs
